+++ src/Event/ObserverFactory.swift
@@ -3,8 +3,7 @@ import Foundation
 public class ObserverFactory {
     public static var currentFactory: ObserverFactory?
 
-    public init() {
-    }
+    public init() {}
 
     public func getObserverForTunnel(tunnel: Tunnel) -> Observer<TunnelEvent>? {
         return nil