+++ src/Opt.swift
@@ -1,18 +1,20 @@
 import Foundation
 
-struct Opt {
-    static let MAXNWTCPSocketReadDataSize = 15000
+public struct Opt {
+    public static var MAXNWTCPSocketReadDataSize = 15000
 
     // This is only used in finding the end of HTTP header (as of now). There is no limit on the length of http header, but Apache set it to 8KB
-    static let MAXNWTCPScanLength = 8912
+    public static var MAXNWTCPScanLength = 8912
 
-    static let DNSFakeIPTTL = 300
+    public static var DNSFakeIPTTL = 300
 
-    static let DNSPendingSessionLifeTime = 10
+    public static var DNSPendingSessionLifeTime = 10
 
-    static let UDPSocketActiveTimeout = 300
+    public static var UDPSocketActiveTimeout = 300
 
-    static let UDPSocketActiveCheckInterval = 60
+    public static var UDPSocketActiveCheckInterval = 60
 
-    static let MAXHTTPContentBlockLength = 10240
+    public static var MAXHTTPContentBlockLength = 10240
+
+    public static var RejectAdapterDefaultDelay = 300
 }